Imágenes y métodos de TILE
- Arreglo sistemas de coordenadas
- Añadidas imágenes a las casillas
- Añadidos métodos de la clase Tile
- Arreglos menores

+++ src/Punto2D.h
@@ -2,6 +2,6 @@
 class Punto2D
 {
 public:
-	int x, z;
+	double x, z;
 };
